Fix incorrect yum output parsing regexp
It was incorrectly expecting a package name to be composed of a-z
characters only.

Parsing "Requires: python2-hawkey >= 0.7.0" resulted in this message
in the log:
Found req packages: [u'python', u'python', u'python', u'python']

But it should be:
Found req packages: (u'python2-hawkey')

That's why the replacement of using list with set as well.